docs: update tabs
